backend/requirements.txt
django
psycopg2
djangorestframework
djangorestframework
djangorestframework-simplejwt
django-otp
django-two-factor-auth
requests==2.31.0
Pillow
dj-database-url
pyotp
redis 
django-redis
celery[redis]
channels
channels_redis
social-auth-app-django
daphne
twilio
django-cors-headers>=3.7.0
aioredis==1.3.1  
channels-redis==3.3.1
qrcode
python-dotenv
